all files except raycast.c norminette friendly, cub3d on branch now matches current version, Makefile updated
