+++ Makefile
@@ -6,7 +6,7 @@
 #    By: oscarmathot <oscarmathot@student.42.fr>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/12/03 17:33:31 by oscarmathot       #+#    #+#              #
-#    Updated: 2024/02/10 15:31:12 by oscarmathot      ###   ########.fr        #
+#    Updated: 2024/03/18 16:14:36 by oscarmathot      ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -15,7 +15,9 @@
 
 NAME 	:= cub3d
 SRC 	:= main.c helpers.c read_map.c file_checks.c initialize.c map_checks.c raycast.c utils.c \
-			debug.c mlx.c player.c mouse.c texture.c cnf.c
+			mlx.c player.c mouse.c texture.c player_helper.c filedata.c check_helper.c cnf.c \
+			edge_verification.c map_checks_helper.c mlx_helper.c stack.c mlx_walls.c rgba.c \
+			texture_init.c
 SUBDIR	:= srcs/
 SRCS	:= $(addprefix $(SUBDIR),$(SRC))
 OBJ 	:= $(SRCS:.c=.o)
