+++ cub3d.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub3d.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: oscar <oscar@student.42.fr>                +#+  +:+       +#+        */
+/*   By: oscarmathot <oscarmathot@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/03 17:28:46 by oscarmathot       #+#    #+#             */
-/*   Updated: 2024/03/17 16:56:28 by oscar            ###   ########.fr       */
+/*   Updated: 2024/03/18 16:20:12 by oscarmathot      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -163,26 +163,13 @@ typedef	struct s_texture_vars
 	int				y_diff;
 	int				end_y;
 	int				screen_height;
-	double				corrected_height;
+	int				corrected_height;
 	int				size;
 	double			magnitude;
 	int				texture_y;
 	int				cur_screen_x;
 }	t_texture_vars;
 
-
-typedef struct s_ray_cast
-{
-  double  current_x;
-    double  current_y;
-    t_point  player_offsets;
-    double  change_wall_size_look;
-    t_point  *good_coord;
-    double  slope;
-    t_point  cur_coord;
-    double  lenght_check;
-
-} t_ray_cast;
 
 void		print_map(char **map);
 bool		ft_isspace(unsigned char c);
